Rename TokenList route to Tokens and cleanup
