+++ packages/wallet/src/App/routes/Tokens/Tokens.less
@@ -1,0 +1,18 @@
+/* Mixins: .StandardPageLayout() */
+@import "~@cosmicdapp/design/theme/layout/mixins.less";
+
+.Tokens {
+  .StandardPageLayout();
+
+  .MainStack {
+    --gap: var(--s4);
+  }
+
+  h1 {
+    margin: 0;
+  }
+
+  .ant-form {
+    margin-top: var(--gap);
+  }
+}
